+++ src/main/java/org/onedatashare/server/controller/endpoint/EndpointBaseController.java
@@ -1,10 +1,16 @@
 package org.onedatashare.server.controller.endpoint;
 
 import org.onedatashare.server.model.core.Stat;
+import org.onedatashare.server.model.error.AuthenticationRequired;
+import org.onedatashare.server.model.error.DuplicateCredentialException;
+import org.onedatashare.server.model.error.ODSAccessDeniedException;
+import org.onedatashare.server.model.error.TokenExpiredException;
 import org.onedatashare.server.model.request.OperationRequestData;
 import org.onedatashare.server.model.request.RequestData;
+import org.onedatashare.server.service.ODSLoggerService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -55,4 +61,26 @@ public abstract class EndpointBaseController {
     protected abstract Mono<Stat> uploadOperation();
     protected abstract Mono<String> downloadOperation(RequestData requestData);
     protected abstract Rendering oauthOperation();
+
+    @ExceptionHandler(AuthenticationRequired.class)
+    public ResponseEntity<String> handle(AuthenticationRequired authenticationRequired) {
+        return new ResponseEntity<>(authenticationRequired.toString(), authenticationRequired.status);
+    }
+
+    @ExceptionHandler(TokenExpiredException.class)
+    public ResponseEntity<String> handle(TokenExpiredException tokenExpiredException) {
+        return new ResponseEntity<>(tokenExpiredException.toString(), tokenExpiredException.status);
+    }
+
+    @ExceptionHandler(ODSAccessDeniedException.class)
+    public ResponseEntity<String> handle(ODSAccessDeniedException ade) {
+        return new ResponseEntity<>("Access Denied Exception", HttpStatus.FORBIDDEN);
+    }
+
+    @ExceptionHandler(DuplicateCredentialException.class)
+    public Rendering handleDup(DuplicateCredentialException dce) {
+        ODSLoggerService.logError(dce.status.toString());
+        return Rendering.redirectTo("/transfer").build();
+    }
+
 }
